// hotelpg/src/app/models/Extra.ts
export interface Extra {
    id_extra?: number;
    nombre?: string;
    fecha?: string;
    estado?: string;
    last_update?: string;
    id_usuario_op?: number;
    id_hotel?: number;
    id_hospedaje?: number;
    cantidad?: string;
    cantidad_cancelados?: string;
    precio_unitario?: string;
}